# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: MoxApp API
  description: |
    High-performance concurrent HTTP load testing tool with DNS timing metrics.
    
    This API provides endpoints for:
    - **Outgoing Traffic Management**: Configure and control external API calls
    - **Incoming Traffic Simulation**: Define mock routes with configurable response patterns
    - **Metrics & Monitoring**: Real-time metrics collection and health monitoring
    - **Scheduler Control**: Pause, resume, or emergency stop the load test
  version: 1.0.0
  contact:
    name: MoxApp Team
  license:
    name: Proprietary
    identifier: LicenseRef-MoxApp
servers:
  - url: __SERVER_URL__
    description: Current server
...
paths:
...
  /api/outgoing/auth-configs:
    get:
      tags:
        - Auth Configs
      summary: List auth configs
      description: Returns all auth configs (credentials masked)
      operationId: listAuthConfigs
      responses:
        '200':
          description: Auth configs list
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  auth_configs:
                    type: object
                    additionalProperties:
                      $ref: '#/components/schemas/AuthConfig'
        '503':
          description: Configuration manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Auth Configs
      summary: Create auth config
      description: Creates a new auth config
      operationId: createAuthConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthConfigRequest'
      responses:
        '201':
          description: Auth config created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  auth_config:
                    $ref: '#/components/schemas/AuthConfig'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Auth config already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Configuration manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/outgoing/auth-configs/{name}:
    get:
      tags:
        - Auth Configs
      summary: Get auth config
      description: Returns a single auth config by name (credentials masked)
      operationId: getAuthConfig
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Auth config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthConfig'
        '404':
          description: Auth config not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Configuration manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Auth Configs
      summary: Update auth config
      description: Updates an existing auth config
      operationId: updateAuthConfig
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthConfigRequest'
      responses:
        '200':
          description: Auth config updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  auth_config:
                    $ref: '#/components/schemas/AuthConfig'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Auth config not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Auth config already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Configuration manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Auth Configs
      summary: Delete auth config
      description: Deletes an auth config by name
      operationId: deleteAuthConfig
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Auth config deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Auth config not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Configuration manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
    AuthConfig:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [none, bearer, api_key, api_key_query, basic, custom_header]
        description:
          type: string
        header_name:
          type: string
        query_param:
          type: string
        env_var:
          type: string
        username_env:
          type: string
        password_env:
          type: string
        token_endpoint:
          $ref: '#/components/schemas/AuthTokenEndpoint'
        refresh_before_expiry:
          type: integer
    AuthConfigRequest:
      type: object
      required:
        - name
        - type
      properties:
        name:
          type: string
        type:
          type: string
          enum: [none, bearer, api_key, api_key_query, basic, custom_header]
        description:
          type: string
        header_name:
          type: string
        query_param:
          type: string
        env_var:
          type: string
        username_env:
          type: string
        password_env:
          type: string
        token_endpoint:
          $ref: '#/components/schemas/AuthTokenEndpoint'
        refresh_before_expiry:
          type: integer
    AuthTokenEndpoint:
      type: object
      properties:
        url:
          type: string
        url_env:
          type: string
        method:
          type: string
        username_env:
          type: string
        password_env:
          type: string
        headers:
          type: object
          additionalProperties:
            type: string
        body:
          description: Any JSON body
        token_path:
          type: string
        expires_path:
          type: string
...